Inroduce a backand and rework condensing threading

+++ collecting/src/fashion_transparency_index.rs
@@ -15,7 +15,7 @@ pub mod data {
 
         /// Score of the company in the Fashion Transparency Index.
         #[serde(rename = "score")]
-        pub score: usize,
+        pub score: i32,
     }
 }
